remove MIT License
Its terseness is very nice, but its text isn’t licensed under anything
by itself. For software, this is fine. But I want *everything* here to
be freely licensed and usable.

The Creative Commons licenses are tried-and-true for cultural and
artistic works, so they’re great for creative works like those that will
be on gabl.ink. They *can* be used for software, as well, but there are
some pitfalls with that.

The Creative Commons licenses are under CC0 1.0, so those are fine as
they are. Using Creative Commons licenses for source code is generally
frowned upon, however, because they explicitly retain patent rights. No
patent rights will ever forseeably apply to gabl.ink, and even if they
ever did I would license them somehow, but many software package
repositories simply don’t accept Creative Commons licenses. So if
anyone wanted to use my code, adapt the file format and structure, etc.
and then publish them to a package repository, they might just… *not be
able to.* And, per my “mission statement” in the previous paragraph,
this doesn’t sound very favorable to me. So I need to dual-license the
code under a good dedicated software license as well.

The MIT License does not expressly grant patent rights, but neither
does it *not,* which is legally acceptable for most people. So the MIT
License is… well, accepted. But I can’t *use* the MIT License. I
theoretically could and then not include it inside the file system,
just providing a link instead, but that’s not great either, because
then the license isn’t included in the file system. Obviously.

Irritatingly, the texts of *most* licenses don’t have a license.
Creative Commons licenses are under CC0, including CC0 itself. The
Unlicense is also under CC0, but that is what is sometimes known as a
“crayon license,” written in such a way that its interpretation in a
court cannot be reliably predicted (i.e. it’s a bit of a risk to use).

Version 2.0 of the Apache License is licensed under Version 2.0 of the
Apache License. It is a “permissive” license, similar to CC BY, which
means that you can distribute and adapt the work with no restrictions—
besides crediting the author and respecting disclaimers and such, of
course. This means that the text of the Apache License 2.0 is, in fact,
eligible for inclusion on gabl.ink. Even better, it contains an explicit
patent grant, so there’s no question in that regard. Hooray! Our
software license problem is solved.

Well, at least 95% solved. The Free Software Foundation and Open Source
Initiative both recognize Apache 2.0 as a free and open-source license,
but the OpenBSD project does not. This is not ideal, but a *single*
caveat (that I’m aware of, anyway) makes this a “good enough” situation
for me. At the very least, I can include the text of Apache 2.0 and
remove it later without breaking my “openness” requirements. The worst
possible scenario is that I have to link to the MIT License and not
include its text.

None of these licenses release *trademark* rights—Creative Commons and
Apache specifically withhold them, in fact. I *am* planning on
trademarking some things, but I’ll allow permissive use of those
trademarks as well. The important thing for “compatibility” with other
projects and use cases is that I don’t believe any of them have a real
problem with trademarks, so we’re probably fine there.

There are two more *potential* licensing problems. The first is that
some trademarks are *not* permissively licensed. For example,
Creative Commons and Apache hold trademarks that *do* have permitted
uses (“licenses”), but not in *all* uses. Creative Commons licenses
have “badges” which are registered trademarks, but they give permission
to use the badges in order to identify the license one is using. The
badges are licensed under CC BY 4.0, so there are no *copyright*
issues; therefore, I can include the image files for the badges, and it
wouldn’t create any copyright issues *or* inherent trademark issues.
Apache also allows some uses of their trademarks; for example, I could
use the Apache License logo as a hyperlink to the Apache License, and
the Apache License logo is licensed under the Apache License, so we’re
good on copyright. What I’m saying is that the logos are freely
distributable and adaptible—just not the trademarks that come with them.
I’ll have to do more research to decide if I’m okay with this or not.

The second is that I plan on including some files using “non-free” file
formats. However, I’m not sure that this is a problem at all, either.
MPEG-4 Part 2 (MP4 video files) is almost entirely expired patent-wise.
The most compatible MP4 codecs, H.264 and AAC, are not, however. Even
with patents, you *can* use the files, but it might *possibly* be
compromising “openness.” Clip Studio Paint work files aren’t patented
at all as far as I’m aware, but the format also isn’t open source. This
probably isn’t a problem either. But that’s what I said before—these are
*potential* problems.

I need to get the heck to sleep now; it’s like 11:20 p.m. right now
(05:20 in UTC, for disambiguation). I hope I never have to write a
commit message as long as this ever again. Good night.