reverse-recursively operate on items in “commit.sh”
Alright, it’s time for another long commit message, but I’ll try to
make it not as long as last time. *shudders*

I’m thinking about how to internationalize stuff. Before, I was
planning on translating the names of all the JSON keys as well as the
content (so “title” would become “titre” in French, for example).
Apps could determine what key names to query from an ID’s language tag.

This would allow users to read the JSON in their preferred language,
not just the generated pages, and I think that would be pretty good.

However, it _would_ make it harder to write programs to work with that
JSON. There are a few ways to make it work, though:

  1. Add translations in a dictionary;
  2. Give the different-language schemas the exact same key order (this
     format would be more difficult to parse);

And the third, and easiest for me:

  3. Don’t!

And just use English. But I don’t want that to be the case, so that’s
not ideal. Oh well. Bye-bye!

(P.S. Also, I’m going to start adding commit trailers.)

Authored-by: Gabriel “gabldotink” <gabl@gabl.ink>
Format: markdown
Signed-off-by: Gabriel “gabldotink” <gabl@gabl.ink>
SPDX-License-Identifier: CC-BY-4.0