moved some styling to comic_page.css
It’s sorta bothering me that generated CSS content (like quotation
marks) aren’t copyable by the reader. There’s a W3C draft spec for an
attribute to change that, but it’s not implemented. I’ll have to keep
checking back on that every once in a while.