some stuff
I’m honestly a bit teed off by how developers are sometimes. “What
possible use case could you have for allowing users to copy what’s
on the screen in an intuitive and reasonable manner?” That’s probably
not the way the W3C has’t added a property for that yet, but it feels
like it somehow.